refactor: src/comments->src/comment
