+++ src/index.ts
@@ -1,7 +1,7 @@
 // This file is the single public-facing API of the entire library.
 
 export * from "./achievement";
-export * from "./comments";
+export * from "./comment";
 export * from "./console";
 export * from "./feed";
 export * from "./game";
